Rework CTE binding: remove CTENode, and bind CommonTableExpressionMap directly instead (https://redirect.github.com/duckdb/duckdb/pull/19351)
The current CTE binding code is rather confusing, as we effectively have
two locations in which CTEs are stored.

* The CTENode, which stores a CTE definition followed by the rest of the
query (potentially containing other CTEs)
* The CommonTableExpressionMap, that was the original location in which
CTEs were stored

Effectively a query like this:

```sql
WITH cte AS (SELECT 42)
SELECT * FROM cte;
```

Would be stored like this:

```
CTENode
   ctename -> "cte"
   query   -> "SELECT 42"
   child   ->
      SelectNode
          CommonTableExpressionMap
          	"cte" -> "SELECT 42"
          query -> "SELECT * FROM cte"
```

As we can see, the CTE is stored twice: once as a CTENode, and once in
the CommonTableExpressionMap. For binding, the CTENode is used. However,
for other purposes, e.g. printing to string, the
CommonTableExpressionMap is used.

This is all rather confusing - and duplicating the CTEs across these two
locations can lead to problems. For example, when reading from legacy
files, the CTENode is not present and must therefore be constructed. We
have special code in a number of places to accomplish this
"re-duplication".

A number of places actually lacked this re-duplication as well, leading
to bugs. For example, when writing the below views using v1.0 and
reading them with v1.4, the CTENode is not correctly inserted and we run
into errors when trying to query the below views.

```sql
-- works, we do re-duplication at the top level
create view v1 as WITH cte AS (SELECT 42)
                          SELECT * FROM cte;

-- does not work, we do not do re-duplication within subqueries
create view v2 as select * from (WITH cte AS (SELECT 42)
                          SELECT * FROM cte);

-- does not work, we do not do re-duplication within CTEs
create view v3 as with cte2 as (WITH cte AS (SELECT 42)
                          SELECT * FROM cte) from cte2;
```


### Changes

This PR gets rid of the CTENode, and moves back to using the
CommonTableExpressionMap for binding. That means that the above query
now looks like this similar to how it used to look:

```
SelectNode
    CommonTableExpressionMap
      "cte" -> "SELECT 42"
    query -> "SELECT * FROM cte"
```

Binding on CTEs is now done by handling any CTEs when binding a
QueryNode using effectively the same binding code as we had when binding
a CTENode - but it is now done explicitly through calling `PrepareCTE`,
followed by binding the actual QueryNode, followed by calling
`FinishCTE`. The code path that is executed should be identical, it just
doesn't use recursion as the previous implementation did.

I actually created a second changeset that [tried to get rid of the
CommonTableExpressionMap instead of the
CTENode](https://github.com/duckdb/duckdb/compare/main...Mytherin:duckdb:removectemap),
but that (1) turned out to be much more complex from a backwards
compatibility perspective as we would need to parse the
`CommonTableExpressionMap` and re-introduce CTENodes, and (2) could lead
to stack overflow issues as we would make much deeper query plans when
we had many CTEs. The approach of removing the CTENode seems to me like
the cleanest solution to the issue.


### Serialization

As this has all happened at the parse-node level, the CTENode does get
serialized to files when writing views with CTEs in DuckDB <= 1.4.0. In
particular, the CTENode is created when writing materialized CTEs (which
as of v1.4.0 is all CTEs). For that reason, the CTENode is kept around
only for backwards compatibility purposes. When deserializing a CTENode,
we deserialize the node and then immediately discard it again. As all
CTEs should have both the CTENode and the CommonTableExpressionMap, this
is sufficient to provide backwards compatibility for binding CTEs.
CTENodes should never be created outside of this temporary pattern, and
support for them has been removed everywhere else.

CC @kryonix 

Upstream-commit: https://github.com/duckdb/duckdb/commit/cb636bc87433541aa326373012baf4afe05f651f