Avoid keeping shared ptr to ClientContext in QueryContext (https://redirect.github.com/duckdb/duckdb/pull/19353)
This fixes a thread sanitizer issue that has been popping up due to
this, see e.g.
[here](https://github.com/Mytherin/duckdb/actions/runs/18417785208/job/52491151141).

The issue is in particular in `UndoBuffer::Cleanup`. Currently, we log
it using the original transaction that created the undo buffer. This is
however not the transaction that is initiating the current operation -
and that connection might have already been destroyed / be in the
process of being destroyed. This leads to a lock inversion (and
potential deadlock) if the client contexts' destroy is triggered in this
location.

In general it seems like this is not the right solution at this location
given that we are passing in a ClientContext that is not actually
operating there at that moment. We should be passing in the actual
client context that is operating.

The main confusion here is that, in general, a Transaction has a
ClientContext associated with it - and the Transaction only exists while
the ClientContext exists. However, because we store transactions after
the client is done with them (in `recently_committed_transactions` and
`old_transactions`) this is no longer the case during that window. I
think this is not very intuitive from a code perspective.

I would propose reworking this so that we are not storing a
`DuckTransaction` in these places, but instead storing only the relevant
components of that transaction (i.e. the `UndoBuffer`). We can then
rework `Transaction` to no longer have a `weak_ptr<ClientContext>` but
to just have a `ClientContext &` - as the transaction will only be alive
while the ClientContext itself is alive and is not kept around past the
lifetime of its parent connection.

Upstream-commit: https://github.com/duckdb/duckdb/commit/2762f1aa72388ad8bbbfe7e2f03089032b8e833d