skip several ci jobs on prs that only bump extensions (https://redirect.github.com/duckdb/duckdb/pull/19249)
When splitting out the '.github/config/out_of_tree_extensions.cmake'
file into 'github/config/extensions/*.cmake' we did not update the
workflow triggers, meaning that simple PRs that only bump extensions
will trigger CI that does not actually do any useful work.

This PR restores that and tries to limit the amount of work that is done
on extension bumps. This is a step towards automating the process of
updating extension versions.

Upstream-commit: https://github.com/duckdb/duckdb/commit/f707ac5f0ad237380153c56275c40f9c9b7d7c20