Parse result for PEG transformer (https://redirect.github.com/duckdb/duckdb/pull/19229)
Follow-up PR to https://github.com/duckdb/duckdb/pull/19181 (I rebased
this branch onto the one in the PR so it should include everything)

This PR introduces (once again) the `ParseResult` concrete syntax tree
structure used by the PEG Transformer.
These `ParseResult`s make use of a `ParseResultAllocator` similar to the
`MatcherAllocator`.

The PR also fixes a (potential) bug related to improperly initializing
the `ParserExtension`. I now initialize all those pointers to `nullptr`
otherwise they can contain garbage value I found.

I reworked the way the PEG Parser is called, it now uses the
`ParserExtension` and `parser_override_function_t` introduced in:
https://github.com/duckdb/duckdb/pull/19126

It includes the structure of the `PEGTransformer` but does not actually
transform any form of statement yet. My plan is to introduce some of the
simpler statements, such as `SET`, `RESET`, `PRAGMA` in a follow-up PR.

Regarding testing this new transformer I was considering adding a new
mode to `ALTERNATIVE_VERIFY` that reruns the query using the
`PEGTransformer` and checks for inconsistent results. An issue I see is
that the list of incorrect results will be huge until a large part of
the transformer is implemented. Should all those tests simply be
excluded or can we run subsets of tests and exclude other tests?

Upstream-commit: https://github.com/duckdb/duckdb/commit/4a221c652ad804599fc0adc534720a6fcd908837