Allow * NOT SIMILAR TO 'pattern' (https://redirect.github.com/duckdb/duckdb/pull/19232)
Fixes https://github.com/duckdblabs/duckdb-internal/issues/5955

Previously the following query would crash stating that the
`StarExpression` must be at the root of the expression.
```sql
create or replace table foo as select 'd' as a, 'f' as b, 'c' as c;
select * not similar to '(a|c)' from foo;
```
This PR fixes that by checking for `OPERATOR_NOT` inside the
`TryTransformStarLike` method and rewriting it to a lambda:
`list_filter(*, lambda __lambda_col: __lambda_col not similar to
'pattern')` if it finds there to be a `NOT regexp_full_match` with a
`FunctionExpression` containing a `StarExpression`. This is similar to
other cases such as is done for `LIKE`.

Upstream-commit: https://github.com/duckdb/duckdb/commit/7bc5d9e0ac8fa29073300458c24856228e9b77d3