Make Binding CTEs lazy, and other CTE binding refactors (https://redirect.github.com/duckdb/duckdb/pull/19372)
This PR makes CTE binding lazy, so that CTEs are not bound eagerly but
are instead bound when first referenced.
This PR also does a number of other binder related clean-up:

* Make Binding members private and use getters
* Remove `CTE_bindings` in the `Binder` - instead only use the
`cte_bindings` in the BindContext
* Allow CTE bindings to have full bind aliases (including schema names)
and use this for recurring CTEs - these are now called `schema:
recurring, table: {cte_name}` instead of `table: recurring. + {ctename}`

CC @kryonix @lnkuiper 

Upstream-commit: https://github.com/duckdb/duckdb/commit/604e5ce24ed809f3c145278ba4da5d892e2ef918