Destroy TaskNotifier prior to calling FinishTask (https://redirect.github.com/duckdb/duckdb/pull/19373)
This fixes a rare race condition when destroying a Connection while
there are active tasks for that connection. When a connection is
destroyed, all tasks for that connection are cleaned up. This is done
through notifications from `Executor::FinishTask`. In the current
implementation, we first notify the executor the task is finished, and
then destroy the TaskNotifier triggering it to notify any registered
ClientContextStates. However, these live in the Connection. By
destroying the TaskNotifier first we ensure the client is still around
at that point.

Upstream-commit: https://github.com/duckdb/duckdb/commit/50f3d6462047362c7b78b6b1cb395904a14da4cb