Avoid calling shared_from_this() and instead use the passed in ClientContext in buffered data (https://redirect.github.com/duckdb/duckdb/pull/19379)
Fixes https://github.com/duckdblabs/duckdb-internal/issues/6234

When running the `ClientContext` destructor, we cancel all pending tasks
that are scheduled for execution for that client. This is required
because tasks generally access the ClientContext all over.

This works just fine - except that we are in a fun situation in a
destructor. While the `ClientContext` is still valid and can be used as
normal - `weak_ptr.lock()` returns `nullptr`. That is because [in order
to construct a valid shared_ptr, we need the object to be alive, and
that is not the case while it is being
destroyed](https://stackoverflow.com/questions/28338978/stdenable-shared-from-this-is-it-allowed-to-call-shared-from-this-in-destru).
As a result, we could rarely run into a `Attempted to dereference
shared_ptr that is NULL` error in the result collection code as that
would call `weak_ptr.lock()` followed by accessing the ClientContext.

This PR fixes that issue by instead using the ClientContext that is
passed in directly.

Upstream-commit: https://github.com/duckdb/duckdb/commit/d1eaf5e700f20806c900ac1816ee26c3e1416647