BUGFIX: Silent failure to write row groups with large lists (https://redirect.github.com/duckdb/duckdb/pull/19376)
When the contents of all lists within a column within a rowgroup are >
int32 max duckdb silently fails to write valid parquet files potentially
causing loss of data.

This fix changes both it to loudly fail with an exception in the future,
and implements a fix which allows writing files in this case, however
the fix goes against a listed comment which I was unsure of the
reasoning behind. I failed to find any requirements in the parquet spec
for it, and am assuming it is for performance reasons, so only go
against the comment when absolutely required.

I added a really, really slow test which is able to replicate the issue,
on the current version of duckdb it fails on the read, but not the write
with:
```
Invalid Error:
Out of buffer
```
However, after just replacing that cast with the checked version it will
instead fail on the write, then with the other changes it no longer
fails at all, but switching to the checked cast seems like good practice
in case there are any other similar issues anywhere else.


It seems that the test was too slow and was getting a SIGTERM (it may
also be running out of RAM, I'm not sure)

Here is the relevant test, it's difficult to run all the time because
operations with large lists in duckdb are generally quite slow,
especially when getting up to 2GB in size:
```
require parquet

statement ok
CREATE TABLE test_10m_arrays AS
SELECT list_resize([]::TINYINT[], 10000000, 1::TINYINT) AS large_array
FROM range(300);

statement ok
COPY (SELECT * FROM test_10m_arrays) TO '__TEST_DIR__/test_10m_arrays.parquet';

statement ok
DROP TABLE test_10m_arrays;

query I
SELECT large_array.len().avg() AS avg_length FROM '__TEST_DIR__/test_10m_arrays.parquet';
----
10000000.000000
```

Upstream-commit: https://github.com/duckdb/duckdb/commit/d13e53bbb666b3f961d5cb029ca939fbefecd802