Fixes for CTE (de)serialization compatibility with older versions (https://redirect.github.com/duckdb/duckdb/pull/19393)
The CTE (de)serialization code in v1.4 has a number of issues:

* It is writing `CTEMaterialize` for CTENodes which is not supported in
older DuckDB versions, causing forwards compatibility to break and older
versions not to be able to read DuckDB files written by v1.4 when they
contain CTEs that are explicitly labeled as `MATERIALIZED` or `NOT
MATERIALIZED`
* It is not correctly de-duplicating CTENodes from the
CommonTableExpressionMap, causing some old CTEs to not be readable
anymore (as explained here -
https://github.com/duckdb/duckdb/pull/19351)

This has all already been fundamentally fixed in main in
https://github.com/duckdb/duckdb/pull/19351. However, in order to also
fix this for following v1.4 versions (v1.4.2) - this PR patches the
serialization code in a lower risk manner. Effectively:

* We no longer write `CTEMaterialize` for CTENodes. Instead, we only
write the CTENode if it should be materialized. Otherwise, we don't
write it to the file.
* When reading a `QueryNode`, we immediately perform "re-duplication" by
extracting CTENodes from the `CommonTableExpressionMap. This fixes an
issue where we were not correctly re-duplicating at all levels.
* When deserializing a CTENode, we perform de-duplication of CTEs within
its child. This fixes an issue where the above re-duplication could
cause the same CTE to appear multiple times depending on from which
version we are de-serializing.

All of this is only necessary for v1.4 - and when we merge v1.4 into
main this code should be deleted.

Upstream-commit: https://github.com/duckdb/duckdb/commit/d8ae68f883fd79a064471dc2b2a2fb03aa7129d8