Use statement for PEG Parser transformer (https://redirect.github.com/duckdb/duckdb/pull/19302)
PR based on: https://github.com/duckdb/duckdb/pull/19181,
https://github.com/duckdb/duckdb/pull/19229

This PR introduces the first `Statement` in the transformer, the
`UseStatement` to select the database and/or schema.
 
Other things in this PR:
- During tokenization we now keep track of the offset and length for
each token
- I added a new config to the `Main` workflow that tests with the
`allow_parser_override_extension` set to `fallback`. This should test
that the new transformer will either (1) parse and transform correctly
and return the correct result or (2) throw an error and let the old
parser/transformer handle the query. So it should pass all tests.
- The `ParserOverrideResult` now stores `ErrorData` rather than the raw
string for better error handling and cleaner error messages.

Upstream-commit: https://github.com/duckdb/duckdb/commit/cf51328b172d4531c942943ef1c47d09ec1f90c9