Fix https://redirect.github.com/duckdb/duckdb/pull/19211: make INSERT OR IGNORE correctly handle multiple constraints (https://redirect.github.com/duckdb/duckdb/pull/19409)
Fixes https://redirect.github.com/duckdb/duckdb/pull/19211

We push a `DISTINCT ON (unique_col)` to filter for unique values within
each column. When dealing with multiple constraints, we need to push
multiple `DISTINCT ON` clauses. This can only happen with `INSERT OR
IGNORE` - because that is the only statement type that supports handling
multiple constraints at the same time.

Upstream-commit: https://github.com/duckdb/duckdb/commit/62a68d6f7b024547c41c36c5239a310b219be289