Fix issue in MetadataManager triggered when doing concurrent reads while checkpointing, and rework concurrent attach / detach test (https://redirect.github.com/duckdb/duckdb/pull/19424)
This PR fixes an issue in the metadata manager when doing concurrent
reads while checkpointing. As we are now lazily reading metadata during
reads, it is possible that we are reading a block in one connection
while another thread is writing out blocks as part of a checkpoint
operation. This previously used the `ConvertToPersistent` callback to
accomplish this. The issue is that `ConvertToPersistent` is destructive
- and takes ownership of the old block in order to avoid re-allocating a
buffer. That only works if the old block is not in active use.

This PR adds a `ConvertToPersistentMode` flag that can be set to
`ConvertToPersistentMode::THREAD_SAFE` which will create a copy of the
old block, allowing the old block to still be safely used while we are
doing the conversion. This could otherwise lead to various different
internal exceptions being thrown in this scenario.

In addition, this PR reworks the concurrent attach / detach test to (1)
split up the task generation from the task execution, and (2) move
everything to C++ classes instead of using global variables within the
test. This allows the test to e.g. be run in a loop trivially.

Upstream-commit: https://github.com/duckdb/duckdb/commit/e3e408271e0a38a5927f04c246bc3e2ad27abac1