Implement `std::allocator` that uses `ArenaAllocator` (https://redirect.github.com/duckdb/duckdb/pull/19412)
This PR implements `arena_stl_allocator`, which implements the
`std::allocator` interface. It allocates using the `ArenaAllocator`.
This can allow us to have many small STL container allocations that live
in the same `ArenaAllocator`, instead of all being tiny individual
allocations.

I've added it in the `PhysicalTopN` operator as a test (as the allocator
to a `vector`), but we should try adding this in various parts of the
code.

CC @taniabogatsch 

Upstream-commit: https://github.com/duckdb/duckdb/commit/f256271bbc8bc1d84f9ca025e6b0423af6f43b6b