Add forwards compatibility tests to CI (https://redirect.github.com/duckdb/duckdb/pull/19432)
Includes https://github.com/duckdb/duckdb/pull/19420

This PR adds forwards compatibility testing to the CI. Example usage of
the script:

```
python3 scripts/test_storage_compatibility.py --versions "1.2.1|1.3.2" --new-unittest build/release/test/unittest
```

The script works by running the unittester (compiled with the latest
version) with the config `test/configs/storage_compatibility.json` for
each test with a `CREATE TABLE` or `CREATE VIEW` statement. The provided
config then runs the test in `--force-storage` mode, but with a fixed
database path (`bwc_storage_test.db`). This then gives us a database
file that has a number of tables / views in it as created in the test.

For each version specified, we then try to read all tables and views
stored within that file. If any of the queries fail, we report an error
**if** the original CLI can successfully query the table / view. The
reason we also do this is that there might be invalid views in the file,
e.g. views that refer to files that have since been deleted.

In order to get the older versions, our install script is used (i.e. we
run `f'curl https://install.duckdb.org | DUCKDB_VERSION=1.3.2 sh` for
each included version).

Upstream-commit: https://github.com/duckdb/duckdb/commit/ff0e9d39e0925a6f31ea88aad48567c5941b12e2