Make cached hashes thread-safe and emit dictionary vectors from Perfect Hash Join (https://redirect.github.com/duckdb/duckdb/pull/19332)
Started from https://redirect.github.com/duckdb/duckdb/pull/19274

This PR makes the cached hashes of string dictionaries thread-safe by
adding a `mutex`, so that we can use `Vector` in parallel. We don't do
this anywhere in the code base, so to also test this functionality, I've
changed the perfect hash join to emit dictionary vectors that have a
size and identifier (just like the string dictionaries that come from
our storage). Emitting these allows further dictionary-based
optimizations during execution, such as our dictionary aggregation and
dictionary functions.

This seemed to give minor speedups in the regression test on my fork,
but it was quite small so it might be just noise.

Upstream-commit: https://github.com/duckdb/duckdb/commit/56cfbb0c2b53fa84b4eb09fb21034357a68195fd