[C-API] Add support for defining config options (settings) (https://redirect.github.com/duckdb/duckdb/pull/19473)
This PR adds support for defining custom configuration options
(settings) from the C-API.

The state of extension/configuration-options/settings and setting
"scopes" in DuckDB are currently in a bit of a mess. I've tried to hide
their differences in the C-API by only introducing a single
`duckdb_config_option_scope` enum that can be used both when defining
the default scope of a config option as well as determine from which
scope a setting was derived when querying a config option.

The only discrepancy is that we currently allow config options to be
declared and set with `SESSION`, `GLOBAL` and an unused
yet-to-be-implemeted transaction-`LOCAL` scope, but in the docs (and
when querying from `duckdb_settings()` we present all `SESSION` local
settings as `LOCAL`. Basically, we in c++ we have both a `SetScope` and
`SettingScope` enum that represent the same things but differently. This
should probably be fixed/changed/alleviated C++-side, but I think it
makes sense to define the enum as the C-API to as `LOCAL/SESSION/GLOBAL`
as that seems to be the way we want all our settings to be defined
eventually (judging from recent work on settings such as
https://github.com/duckdb/duckdb/pull/2609)

Since `ExtensionOption`'s currently don't store any extra state
somewhere, and we are still figuring out extension lifetimes, this PR
does not implement a way to set a callback to be notified when a config
option changes. That will be done in a follow-up PR.

Upstream-commit: https://github.com/duckdb/duckdb/commit/6a144a5b7ea600ce2d4acdcb9faf4d1d5d917970