`GEOMETRY` Rework: Part 3 - Filter pushdown (https://redirect.github.com/duckdb/duckdb/pull/19439)
This is a followup PR that builds on top of
https://github.com/duckdb/duckdb/pull/19203. Please have a look at
https://github.com/duckdb/duckdb/pull/19136 for the context behind this
PR.

In https://github.com/duckdb/duckdb/pull/19203 we added support for
storing geometry statistics. In this PR we now add a `&&`
bounding-box-intersection binary operator that when used in a filter can
be pushed down into storage and prune row-groups based on these new
geometry statistics.

Geometry-filter pushdown works for our own storage format, as well as
parquet (if the parquet file contains parquet-native geometry stats).
I've also cleaned up the parquet extension further and removed all
spatial-dependent code, so that all geoparquet related functionality now
works without requiring spatial to be loaded, which simplifies both the
read and write-path for geometries significantly. It also enables us to
run the geoparquet tests in CI when spatial isn't available.

I've also added a couple of basic geometry scalar-functions to convert
to/from WKB and WKT. These will be more fleshed out in the future (to
e.g. handle big-endian WKB and specify WKT precision).

Upstream-commit: https://github.com/duckdb/duckdb/commit/f1f9fc1bec631d44222cd2c4c246a02d5532368f