Fix issue https://redirect.github.com/duckdb/duckdb/pull/17941: Cryptic error message when dropping a column part of a unique constraint.  (https://redirect.github.com/duckdb/duckdb/pull/19465)
When originally dropping a column part of a multi-line UNIQUE restraint,
the error message would say there is no column X in the table. **This is
not accurate.** So the error message has been fixed from that to _Cannot
drop column X because it is referenced in unique constraint
UNIQUE(**<unique_constraint>**)_.

**Changes:**
- updated RemoveColumn() to check multi-column UNIQUE constraints
(through columns vector rather than hasIndex())
- added better error message
- added test cases to check for multi-column UNIQUE constraint
scenarios.

Upstream-commit: https://github.com/duckdb/duckdb/commit/5a2ada5ebd1e620338e783527ad3d790ec66c757