Fix bug initializing std::vector for column names in ADBC API (https://redirect.github.com/duckdb/duckdb/pull/19444)
Passing a size to the constructor of `std::vector` initializes not the
capacity but the size. When elements are subsequently pushed or placed
onto the back of the vector, the vector will grow. If the growth goes
past the capacity, a vector of strings will be reallocated in a way that
may invalidate pointers previously gotten from those strings.

Resolves https://redirect.github.com/duckdb/duckdb/pull/19443.

There may be a cleaner way to express the intent here. I went with a
minimal change. I didn't add a test case because there was an existing
test case that was already failing on Windows.

Upstream-commit: https://github.com/duckdb/duckdb/commit/2eab381739cc18e4355c9d625a273e098debbb7b