Add test that either 'latest' or 'vX.Y.Z' are supported STORAGE_VERSIONs (https://redirect.github.com/duckdb/duckdb/pull/19527)
Connected to https://github.com/duckdb/duckdb/pull/19525, adds a test
that would have triggered.

That test is not build when actually building releases, so that's not
fool-proof, but I think adding this in is helpful.

Tested locally to behave as intended both on dev commit (success) and
tag (fails, fixed via linked PR).

Upstream-commit: https://github.com/duckdb/duckdb/commit/5a6521ca7e744205e4c3b67cab8708e2df87073b