Convert CLI to use C++ API, and remove sqlite API wrappers (https://redirect.github.com/duckdb/duckdb/pull/19536)
When we initially implemented the CLI, we used the SQLite shell code
"as-is" and implemented SQLite's C API to allow it to use DuckDB. As
time has moved on, we have made big changes to the shell, and have
already partially started moving it towards our own code written in C++
(https://github.com/duckdb/duckdb/pull/14473,
https://github.com/duckdb/duckdb/pull/14503,
https://github.com/duckdb/duckdb/pull/14483). However, the shell would
still use the SQLite API wrappers we had made to call into DuckDB.

This PR cleans that up, and moves the shell towards using our C++ API
directly instead. This simplifies a bunch of the code, and should make
it easier to work on improving the shell in the future.

Upstream-commit: https://github.com/duckdb/duckdb/commit/371121cc2480d3d569e8543d27d21b74990eb0f5