Delete Statement and expression associativity for PEG Transformer (https://redirect.github.com/duckdb/duckdb/pull/19507)
Based on https://github.com/duckdb/duckdb/pull/19460 (that should be
merged first)

Introduce new statements: 
- TransactionStatement
- CallStatement
- CheckpointStatement
- DeallocateStatement
- DeleteStatement

I also reworked the expressions in the grammar. It now respects the
associativity rules as specified by the PostgreSQL grammar, see
[here](https://www.postgresql.org/docs/16/sql-syntax-lexical.html#SQL-PRECEDENCE).
It makes the grammar around the expressions a bit more difficult to
follow, but we need this to get the right order of associativity.

This PR also includes a lot of new expressions, and the `TableRef`, such
as `TableFunctionRef`, `JoinRef`. It doesn't include the
`SubqueryExpression` yet, because that would need basically the entire
`SelectStatement`. I'll include that in a follow-up PR.

Upstream-commit: https://github.com/duckdb/duckdb/commit/9b74cb8f92e04b06aa369d5def3d62a6049dc27c