Adds struct variant to `regexp_extract_all` (https://redirect.github.com/duckdb/duckdb/pull/19078)
Duckdb has a really neat variant of `regexp_extract` which allows
providing column names then is able to export each of the capture groups
after a single regex match, this really neat feature wasn't supported
when extracting all results.

There isn't an easy way to do this otherwise, especially without at
least one more call to match.

This adds that functionality and tries to share some code for parameter
parsing with `regexp_extract`.

I tried to add a variety of tests, but with stuff like this there are
quite a few cases so I'm very open to adding any more.

I encountered this while trying to do some processing from a generic
`read_text` function, I also ran into wanting to have a `regexp_instr`
function like snowflake and others.

It is possible to add support for a variable pattern, but it makes the
logic a bit more complicated, there are a few options for how to handle
differing number of capture groups, not supporting it I believe is
consistent with the `regexp_extract` function.

Upstream-commit: https://github.com/duckdb/duckdb/commit/3282b1db73fbd372177638f543d18c32035ba1cb