Bugfixes: Parquet JSON+DELTA_LENGTH_BYTE_ARRAY and sorting iterator (https://redirect.github.com/duckdb/duckdb/pull/19556)
This PR fixes an issue introduced in v1.4.1 with the Parquet reader when
combining a `JSON` column with `DELTA_LENGTH_BYTE_ARRAY` encoding. The
issue was caused by trying to validate an entire block of strings in one
go, which is OK for UTF-8, but for JSON. This PR makes it so we validate
individual strings if the column has `JSON` type.

Fixes https://github.com/duckdb/duckdb/issues/19366

This PR also fixes an issue with the new sorting code, which had an
error in the calculation of subtraction under modulo. I've fixed this,
and unified the code for `InMemoryBlockIteratorState` and
`ExternalBlockIteratorState` with some templating, so now the erroneous
calculation should be gone from both state types.

Fixes https://github.com/duckdb/duckdb/issues/19498

Upstream-commit: https://github.com/duckdb/duckdb/commit/1cba2e741b6622f5be156c061478a6fa66c0f819