Release relevant tests to still be run on all builds (https://redirect.github.com/duckdb/duckdb/pull/19559)
I would propose, at least for the Linux builds, to add back a minimal
amount of tests also on release builds.

They will ensure at a minimum that:
* for a given release, the corresponding storage_version is valid
* for a minor release, that the corresponding name has been set

There are more tests that we might consider basic enough AND connected
to behaviour specific of a release that we might want to add to the
`release` tag.

Fixes https://github.com/duckdb/duckdb/issues/19354 (together with
https://github.com/duckdb/duckdb/pull/19525 that actually added the
name).

Note that given the current release process happens in advance, eventual
test failure are annoying but not fatal, but they will require changes
to code. I am not sure if it's worth having a `keep_going_in_all_cases`
option, basically turning the boolean into a set, but I think it can be
done when need arise.

Upstream-commit: https://github.com/duckdb/duckdb/commit/b68a1696de1a603b59e39efc25da7fc2826a3135