Custom Order Row Group Scans (https://redirect.github.com/duckdb/duckdb/pull/19537)
This PR introduces the ability to scan row groups in a custom order. For
now, we use this to help the TopN operator by re-ordering the row groups
by the row group statistics on the primary sort column. Depending on the
sort order, this comes in two modes:
- `ORDER BY x[, y, z, ...] ASC`: Scan in *ascending* order of
*min*-statistics on `x`
- `ORDER BY x[, y, z, ...] DESC`: Scan in *descending* oder of
*max*-statistics on `x`

For simple orderby-limit queries with limit `N`, we can even guarantee
that we have to scan at most `N` row groups (and therefore only exclude
all others from the scan). For most simple queries of that kind and
depending on the min/max statistics of the next-in-line row group, we
could even restrict the scan to only read a single row group. This would
be part of a follow-up though.

## Performance
To observe the performance impact of this optimization, let's look at
our very favorite time-series benchmark suite data set with ~100M rows
and the following query, which returns the last 100 measurements:
```sql
SELECT * FROM cpu ORDER BY time DESC LIMIT 100;
```
 
As the data is ordered by time ASC, this query is much slower than it
needs to be. Even though the TopN operator continuously updates the
dynamic table filter, that filter never excludes any row groups, and
therefore we scan the whole table.

By re-ordering the row groups, we can exclude most of them, which
eventually leads to nice performance improvements:

|          | Nightly | This PR | Improvement |
| -------- | ------- | ------- | ----------- |
| Cold Run | 0.7s    | 0.02s   | **35x**     |
| Hot Run  | 0.2s    | 0.008s  | **25x**     |

Upstream-commit: https://github.com/duckdb/duckdb/commit/ead3ca87f790fd7e0b1e3efdaf74c727aae47473