fix inconsistent behavior in remote read_file/blob, and prevent union… (https://redirect.github.com/duckdb/duckdb/pull/19531)
Closes https://github.com/duckdb/duckdb-fuzzer/issues/4208
Closes https://github.com/duckdb/duckdb/issues/19090

Our remote filesystem doesn't actually check that files exist when
"globbing" a non-glob pattern. Now we check that the file exists in the
read_blob/text function even if we just access the file name.

Diff is a bit bigger cause I also moved a bunch of templated stuff into
the cpp file.

Upstream-commit: https://github.com/duckdb/duckdb/commit/be0142d4ee0385262520ae2488e8dd11ac213735