Fix edge case in uncompressed validity scan with offset and fix off-by-one in ArrayColumnData::Select (https://redirect.github.com/duckdb/duckdb/pull/19567)
This PR fixes a off-by-one in the consecutive-array-scan optimization
implemented in https://github.com/duckdb/duckdb/pull/16356 as well as an
edge case in our uncompressed validity data scan.

Fixes https://github.com/duckdb/duckdb/issues/19377

I can't figure out how to write a test for this, it seems like no matter
what I do I'm unable to replicate the same storage characteristics as
the database file provided in the issue above.

In the repro we do a scan+skip+scan, where part of the first
`validity_t` in the second scan contains a bunch of zeroes at the
positions "before" the scan window that remain even after shifting. I've
solved it by setting all lower bits up to `result_idx` in the first
`validity_t` we scan, but not sure if this is the most elegant solution.

Strangely enough If we remove all bitwise logic and just do the same
"fall-back" logic as ifdef:ed for `VECTOR_SIZE < 128` it all works
though, so the issue has to be part of the bit-manipulation.

Upstream-commit: https://github.com/duckdb/duckdb/commit/9ea6e07a290db878c9da097d407b3a866c43c8e0