Wal index deletes (https://redirect.github.com/duckdb/duckdb/pull/19477)
This adds support for buffering and replaying Index delete operations
for WAL replay. During WAL replay, index operations are buffered since
the Indexes are not bound yet. During Index binding, the buffered
operations are applied to the Index. UnboundIndex is modified to support
buffering delete operations on top of inserts.
BoundIndex::ApplyBufferedAppends is changed to a
BoundIndex::ApplyBufferedReplays which supports replaying both inserts
and deletes.

Documentation along relevant code paths is added which clarifies the
ordering of mapped_column_ids and the index_chunks being buffered.
Before, the mapping was any order since it was only coming from Index
insert paths. Now, buffering can come from both insert and delete paths,
so both need to make sure to buffer index chunks and the mappings in the
same order, (which is just in sorted order of the physical Index column
IDs).

There is also a bug fix for buffering index data on a table with
generated columns, since the table chunk being created for replaying
buffered operations contained all column types previously, including
generated columns, whereas now it only contains the physical column
layout which is needed for index operations. (ART Index operations take
a chunk of data with only the index columns containing any data, and the
non-Indexed columns are empty).

A catch block is added to Transaction CleanupState::Flush which was
silently throwing away any failures (which caught this WAL replay in the
first place). Also, some test coverage for ART duplicate rowids and a
LookupInLeaf function was added which allows searching for a rowid in a
Leaf that is either inlined, or a gate node to a nested ART.

@taniabogatsch 

Upstream-commit: https://github.com/duckdb/duckdb/commit/407720a34804f0da61d5ba6645c3c44ec6ddf0d8