Python SQLLogicTest - show results in a consistent order (https://redirect.github.com/duckdb/duckdb/pull/19562)
Fixes: https://github.com/duckdblabs/duckdb-internal/issues/5557

This PR fixes a bug in the nightly tests related to the Python
implementation of SQLLogicTest. Previously, when running the tests using
Python, the result order could vary, causing hash mismatches and test
failures. Changed the test to ensure the results are in consistent
order, irrespective of how we run it.

I am not sure if Python SQLLogicTest is being used at the moment, but I
am pushing the fix anyway as it can cause problems in the future.

Upstream-commit: https://github.com/duckdb/duckdb/commit/cb521b25eb557e5f8a043f13938feb72e630e69d