Avoid eagerly resolving the next on-disk pointer in the MetadataReader, as that pointer might not always be valid (https://redirect.github.com/duckdb/duckdb/pull/19588)
When enabling the new [experimental metadata
re-use](https://github.com/duckdb/duckdb/pull/18395), it is possible for
metadata of *some* row groups to be re-used. This can cause linked lists
of metadata blocks to contain invalid references.

For example, when writing a bunch of row groups, we might get this
layout:

```
METADATA BLOCK 1
ROW GROUP 1
ROW GROUP 2 (pt 1)
NEXT BLOCK: 2

->

METADATA BLOCK 2
ROW GROUP 2 (pt 2)
ROW GROUP 3
```

Metadata is stored in a linked list (block 1 -> block 2) - but we don't
need to traverse this linked list fully. We store pointers to individual
row groups, and can start reading from their position.

Now suppose we re-use metadata of `ROW GROUP 1`, but not of the other
row groups (because e.g. they have been updated / changed). Since this
is fully contained in `METADATA BLOCK 1`, we can garbage collect
`METADATA BLOCK 2`, leaving the following metadata block:


```
METADATA BLOCK 1
ROW GROUP 1
ROW GROUP 2 (pt 1)
NEXT BLOCK: 2
```

Now we can safely read this block and read the metadata for `ROW GROUP
1`, **however**, this block contains a reference to a metadata block
that is no longer valid and might have been garbage collected. This
revealed a problem in the `MetadataReader`. In the current
implementation of the `MetadataReader` - when pointing it towards a
block, it would eagerly try to figure out the metadata location of *the
next block*. This is normally not a problem, however, with these invalid
chains, we might try to resolve a block that has been freed up already -
causing an internal exception to trigger:

```
Failed to load metadata pointer (id %llu, idx %llu, ptr %llu)
```

This PR resolves the issue by making the MetadataReader lazy. Instead of
eagerly resolving the next pointer, we only do this when it is actually
required.

Upstream-commit: https://github.com/duckdb/duckdb/commit/d643cefe13de6873f6fb0ecc0bca1c14111cde11